+++ Assets/AccelByte/Core/UnityWebRequestExtension.cs
@@ -4,7 +4,6 @@
 
 using System;
 using System.IO;
-using System.Reflection;
 using System.Text;
 using UnityEngine.Networking;