+++ Assets/AccelByte/Server/ServerEcommerceApi.cs
@@ -3,10 +3,8 @@
 // and restrictions contact your company contract manager.
 
 using System.Collections;
-using System.Collections.Generic;
 using AccelByte.Core;
 using AccelByte.Models;
-using UnityEngine;
 using UnityEngine.Assertions;
 
 namespace AccelByte.Server