+++ Assets/AccelByte/Server/ServerStatisticApi.cs
@@ -6,7 +6,6 @@ using System.Collections;
 using System.Collections.Generic;
 using AccelByte.Core;
 using AccelByte.Models;
-using UnityEngine;
 using UnityEngine.Assertions;
 
 namespace AccelByte.Server